// frontend/src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        CharClass: "digit" | "lowercase" | "uppercase" | "symbol";
// ...
        PasswordPolicyResponse: {
            /**
             * @description 最小の文字数。
             * @example 4
             */
            minLength: number;
            /** @description 少なくとも1文字ずつ含める必要がある文字種。空なら文字種は問わない。 */
            requiredClasses: components["schemas"]["CharClass"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}